modify enable create setting.
